Load YAML from external files with no summary/description. Fixes #93 (#160)
* Load YAML from external files with no summary/description. Fixes #93

If there's no '---' in the text, but it's from an external file, assume
it's all YAML, but if it's not from an external file, assume there is no
YAML (i.e., it's all docstring).

* Added some test cases to the examples.

But they are duplicates in the examples, and may not be helpful.  It
wasn't clear how best to get these into the tests otherwise.